Enable auth in low level API
Add support for a header of the form

    Authorization: Bearer <token>

which will be sent with every request. The token is to be obtained from
some login page, or with the api call that is also enabled, although
left purposefully manual (since we don't want people to write their
passwords anywhere).

The header will enable actions that require some privileges (which will
include requesting an histogram). The admin "scope" allows the requested
token to perform additional tasks related to managing users (the user
requesting it must in turn be admin).